این هماهنگ با فرانتد کار م یکند ولی در نمایش مقدار موجودی کالا به صورت اعشاری مشکل دارد

+++ src/main/java/com/developer/springOracle3/entity/Production.java
@@ -24,7 +24,7 @@ public class Production {
     @Column
     private Integer meterPr;
     @Transient
-    private  Integer remainMeter;
+    private Double remainMeter;
     @Column
     private Date tarikh;
